generate and export CollectionResponses
